+++ packages/gestalt/src/Masonry.tsx
@@ -17,6 +17,7 @@ import {
   ColumnSpanConfig,
   ModulePositioningConfig,
   MULTI_COL_ITEMS_MEASURE_BATCH_SIZE,
+  ResponsiveModuleConfig,
 } from './Masonry/multiColumnLayout';
 import ScrollContainer from './Masonry/ScrollContainer';
 import { getElementHeight, getRelativeScrollTop, getScrollPos } from './Masonry/scrollUtils';
@@ -140,6 +141,19 @@ type Props<T> = {
    * This is an experimental prop and may be removed or changed in the future.
    */
   _getColumnSpanConfig?: (item: T) => ColumnSpanConfig;
+  /**
+   * Experimental prop to define the minimum and maximum limit a flexible width module could span.
+   * This used to enable multi-column flexible width support ONLY ON SECOND ITEM OF THE ARRAY OF ITEMS.
+   * Also, for the flexible width to work, the _getColumnSpanConfig prop should be set.
+   * _getResponsiveModuleConfigForSecondItem is a function that takes an individual grid item as an input and returns a ResponsiveModuleConfig.
+   * ResponsiveModuleConfig can be one of the following:
+   * - A number, which indicates a static number of columns the item should span
+   * - An object, which sets the minimum and maximum limits a multi-column item could span filling the empty columns in the first row of the grid (flexible width).
+   * - Undefined, which is used to indicate that this prop isn't set.
+   *
+   * This is an experimental prop and may be removed or changed in the future.
+   */
+  _getResponsiveModuleConfigForSecondItem?: (item: T) => ResponsiveModuleConfig;
   /**
    * Experimental flag to enable dynamic heights on items. This only works if multi column items are enabled.
    */
@@ -577,6 +591,7 @@ export default class Masonry<T> extends ReactComponent<Props<T>, State<T>> {
       scrollContainer,
       _logTwoColWhitespace,
       _getColumnSpanConfig,
+      _getResponsiveModuleConfigForSecondItem,
       _getModulePositioningConfig,
     } = this.props;
     const { gutter, hasPendingMeasurements, measurementStore, width } = this.state;
@@ -593,6 +608,7 @@ export default class Masonry<T> extends ReactComponent<Props<T>, State<T>> {
       minCols,
       width,
       _getColumnSpanConfig,
+      _getResponsiveModuleConfigForSecondItem,
       _logTwoColWhitespace,
       _getModulePositioningConfig,
     });