+++ packages/gestalt/src/Masonry/defaultLayout.test.ts
@@ -34,6 +34,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       layout: 'basic',
       minCols: 2,
       rawItemCount: items.length,
+      originalItems: items,
       width: 8000,
       _getColumnSpanConfig,
     });
@@ -66,6 +67,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       layout: 'basic',
       minCols: 2,
       rawItemCount: items.length,
+      originalItems: items,
       width: 8000,
       _getColumnSpanConfig,
     });
@@ -99,6 +101,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       layout: 'basicCentered',
       minCols: 2,
       rawItemCount: items.length,
+      originalItems: items,
       width: 8000,
       _getColumnSpanConfig,
     });
@@ -132,6 +135,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       layout: 'basic',
       minCols: 2,
       rawItemCount: items.length,
+      originalItems: items,
       width: 8000,
       _getColumnSpanConfig,
     });
@@ -164,6 +168,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       positionCache,
       layout: 'basic',
       rawItemCount: items.length,
+      originalItems: items,
       width: 501,
       _getColumnSpanConfig,
     });
@@ -196,6 +201,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       positionCache,
       layout: 'basic',
       rawItemCount: items.length,
+      originalItems: items,
       width: 200,
       _getColumnSpanConfig,
     });
@@ -230,6 +236,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
         gutter: 0,
         layout: align === 'center' ? 'basicCentered' : 'basic',
         rawItemCount: items.length,
+        originalItems: items,
         width: 1000,
         _getColumnSpanConfig,
       })(items);
@@ -279,6 +286,7 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       layout: 'basic',
       minCols: 2,
       rawItemCount: items.length,
+      originalItems: items,
       width: 1000,
       _getColumnSpanConfig,
     });