+++ packages/gestalt/src/Masonry/fullWidthLayout.test.ts
@@ -35,6 +35,7 @@ describe.each([undefined, getColumnSpanConfig])(
         idealColumnWidth: 240,
         minCols: 2,
         width: 1000,
+        originalItems: items,
         _getColumnSpanConfig,
       });
       expect(layout(items)).toEqual([
@@ -70,6 +71,7 @@ describe.each([undefined, getColumnSpanConfig])(
         idealColumnWidth: 240,
         minCols: 2,
         width: 1000,
+        originalItems: items,
         _getColumnSpanConfig,
       });
       expect(
@@ -104,6 +106,7 @@ describe.each([undefined, getColumnSpanConfig])(
         idealColumnWidth: 240,
         minCols: 2,
         width: 1000,
+        originalItems: items,
         _getColumnSpanConfig,
       });