+++ packages/gestalt/src/Masonry/multiColumnLayout.ts
@@ -21,6 +21,7 @@ export type ModulePositioningConfig = {
   whitespaceThreshold?: number; // "Good enough" whitespace number when positioning a module
   iterationsLimit?: number;
 };
+export type ResponsiveModuleConfig = number | { 'min': number; 'max': number } | undefined;
 
 // maps the number of columns to a grid breakpoint
 // sm: 2 columns
@@ -66,15 +67,57 @@ export function getColumnSpanFromGridSize(
   return columnSpanConfig[gridSize] ?? 1;
 }
 
+function getColumnSpanFromResponsiveModuleConfig(
+  columnCount: number,
+  firstItemColumnSpan: number,
+  responsiveModuleConfig: ResponsiveModuleConfig,
+): number {
+  if (typeof responsiveModuleConfig === 'number') {
+    return responsiveModuleConfig;
+  }
+  if (responsiveModuleConfig) {
+    const columnSpan = Math.max(
+      responsiveModuleConfig.min,
+      Math.min(responsiveModuleConfig.max, columnCount - firstItemColumnSpan),
+    );
+    return columnSpan;
+  }
+  return 1;
+}
+
 export function calculateActualColumnSpan<T>(props: {
   columnCount: number;
+  firstItem: T;
+  isFlexibleWidthItem: boolean;
   item: T;
+  responsiveModuleConfigForSecondItem: ResponsiveModuleConfig;
   _getColumnSpanConfig: (item: T) => ColumnSpanConfig;
 }): number {
-  const { columnCount, item, _getColumnSpanConfig } = props;
+  const {
+    columnCount,
+    item,
+    firstItem,
+    isFlexibleWidthItem,
+    _getColumnSpanConfig,
+    responsiveModuleConfigForSecondItem,
+  } = props;
   const columnSpanConfig = _getColumnSpanConfig(item);
   const gridSize = columnCountToGridSize(columnCount);
-  const columnSpan = getColumnSpanFromGridSize(columnSpanConfig, gridSize);
+  // By design, a flexible width module should always be placed at the second position in the grid
+  let columnSpan = getColumnSpanFromGridSize(columnSpanConfig, gridSize);
+
+  if (isFlexibleWidthItem) {
+    const firstItemColumnSpanConfig = _getColumnSpanConfig(firstItem);
+    const firstItemColumnSpan = getColumnSpanFromGridSize(firstItemColumnSpanConfig, gridSize);
+    const responsiveModuleColumnSpan = getColumnSpanFromResponsiveModuleConfig(
+      columnCount,
+      firstItemColumnSpan,
+      responsiveModuleConfigForSecondItem,
+    );
+
+    columnSpan = responsiveModuleColumnSpan;
+  }
+
   // a multi column item can never span more columns than there are in the grid
   return Math.min(columnSpan, columnCount);
 }
@@ -162,19 +205,25 @@ function calculateSplitIndex({
 
 export function initializeHeightsArray<T>({
   centerOffset,
+  checkIsFlexibleWidthItem,
   columnCount,
   columnWidthAndGutter,
+  firstItem,
   gutter,
   items,
   positionCache,
+  responsiveModuleConfigForSecondItem,
   _getColumnSpanConfig,
 }: {
   centerOffset: number;
+  checkIsFlexibleWidthItem: (item: T) => boolean;
   columnCount: number;
   columnWidthAndGutter: number;
+  firstItem: T;
   gutter: number;
   items: ReadonlyArray<T>;
   positionCache: Cache<T, Position> | null | undefined;
+  responsiveModuleConfigForSecondItem: ResponsiveModuleConfig;
   _getColumnSpanConfig: (item: T) => ColumnSpanConfig;
 }): ReadonlyArray<number> {
   const heights = new Array<number>(columnCount).fill(0);
@@ -184,7 +233,14 @@ export function initializeHeightsArray<T>({
       // we do Math.round here because both position.left and columnWidthAndGutter can be floating point numbers
       // in the case of fullWidthLayout (i.e. fluid grid)
       const col = Math.round((position.left - centerOffset) / columnWidthAndGutter);
-      const columnSpan = calculateActualColumnSpan({ columnCount, item, _getColumnSpanConfig });
+      const columnSpan = calculateActualColumnSpan({
+        columnCount,
+        firstItem,
+        isFlexibleWidthItem: checkIsFlexibleWidthItem(item),
+        item,
+        responsiveModuleConfigForSecondItem,
+        _getColumnSpanConfig,
+      });
       // the height of the column is just the sum of the top and height of the item
       const absoluteHeight = position.top + position.height + gutter;
       for (let i = col; i < col + columnSpan; i += 1) {
@@ -503,17 +559,22 @@ function getGraphPositions<T>({
 
 function getPositionsWithMultiColumnItem<T>({
   multiColumnItem,
+  checkIsFlexibleWidthItem,
+  firstItem,
   itemsToPosition,
   heights,
   prevPositions,
   columnCount,
   logWhitespace,
+  responsiveModuleConfigForSecondItem,
   _getColumnSpanConfig,
   _getModulePositioningConfig,
   ...commonGetPositionArgs
 }: {
   multiColumnItem: T;
   itemsToPosition: ReadonlyArray<T>;
+  checkIsFlexibleWidthItem: (item: T) => boolean;
+  firstItem: T;
   heights: ReadonlyArray<number>;
   prevPositions: ReadonlyArray<{
     item: T;
@@ -531,6 +592,7 @@ function getPositionsWithMultiColumnItem<T>({
   gutter: number;
   measurementCache: Cache<T, number>;
   positionCache: Cache<T, Position>;
+  responsiveModuleConfigForSecondItem: ResponsiveModuleConfig;
   _getColumnSpanConfig: (item: T) => ColumnSpanConfig;
   _getModulePositioningConfig?: (gridSize: number, moduleSize: number) => ModulePositioningConfig;
 }): {
@@ -545,7 +607,15 @@ function getPositionsWithMultiColumnItem<T>({
   // This is the index inside the items to position array
   const multiColumnIndex = itemsToPosition.indexOf(multiColumnItem);
   const oneColumnItems = itemsToPosition.filter(
-    (item) => calculateActualColumnSpan({ columnCount, item, _getColumnSpanConfig }) === 1,
+    (item) =>
+      calculateActualColumnSpan({
+        columnCount,
+        firstItem,
+        isFlexibleWidthItem: checkIsFlexibleWidthItem(item),
+        item,
+        responsiveModuleConfigForSecondItem,
+        _getColumnSpanConfig,
+      }) === 1,
   );
 
   // The empty columns can be different from columnCount if there are
@@ -554,7 +624,10 @@ function getPositionsWithMultiColumnItem<T>({
 
   const multiColumnItemColumnSpan = calculateActualColumnSpan({
     columnCount,
+    firstItem,
+    isFlexibleWidthItem: checkIsFlexibleWidthItem(multiColumnItem),
     item: multiColumnItem,
+    responsiveModuleConfigForSecondItem,
     _getColumnSpanConfig,
   });
 
@@ -670,8 +743,10 @@ const multiColumnLayout = <T>({
   logWhitespace,
   measurementCache,
   positionCache,
+  originalItems,
   _getColumnSpanConfig,
   _getModulePositioningConfig,
+  _getResponsiveModuleConfigForSecondItem,
 }: {
   items: ReadonlyArray<T>;
   gutter?: number;
@@ -685,12 +760,27 @@ const multiColumnLayout = <T>({
     numberOfIterations: number,
     columnSpan: number,
   ) => void;
+  originalItems: ReadonlyArray<T>;
   _getColumnSpanConfig: (item: T) => ColumnSpanConfig;
   _getModulePositioningConfig?: (gridSize: number, moduleSize: number) => ModulePositioningConfig;
+  _getResponsiveModuleConfigForSecondItem: (item: T) => ResponsiveModuleConfig;
 }): ReadonlyArray<Position> => {
+  const firstItem = originalItems[0]!;
+  const secondItem = originalItems[1]!;
+  const responsiveModuleConfigForSecondItem = _getResponsiveModuleConfigForSecondItem(secondItem);
+  const checkIsFlexibleWidthItem = (item: T) =>
+    !!responsiveModuleConfigForSecondItem && item === secondItem;
+
   if (!items.every((item) => measurementCache.has(item))) {
     return items.map((item) => {
-      const itemColumnSpan = calculateActualColumnSpan({ columnCount, item, _getColumnSpanConfig });
+      const itemColumnSpan = calculateActualColumnSpan({
+        columnCount,
+        firstItem,
+        isFlexibleWidthItem: checkIsFlexibleWidthItem(item),
+        item,
+        responsiveModuleConfigForSecondItem,
+        _getColumnSpanConfig,
+      });
       if (itemColumnSpan > 1) {
         const columnSpan = Math.min(itemColumnSpan, columnCount);
         return offscreen(columnWidth * columnSpan + gutter * (columnSpan - 1));
@@ -704,11 +794,14 @@ const multiColumnLayout = <T>({
   // the total height of each column
   const heights = initializeHeightsArray({
     centerOffset,
+    checkIsFlexibleWidthItem,
     columnCount,
     columnWidthAndGutter,
+    firstItem,
     gutter,
     items,
     positionCache,
+    responsiveModuleConfigForSecondItem,
     _getColumnSpanConfig,
   });
 
@@ -716,7 +809,15 @@ const multiColumnLayout = <T>({
   const itemsWithoutPositions = items.filter((item) => !positionCache?.has(item));
 
   const multiColumnItems = itemsWithoutPositions.filter(
-    (item) => calculateActualColumnSpan({ columnCount, item, _getColumnSpanConfig }) > 1,
+    (item) =>
+      calculateActualColumnSpan({
+        columnCount,
+        firstItem,
+        isFlexibleWidthItem: checkIsFlexibleWidthItem(item),
+        item,
+        responsiveModuleConfigForSecondItem,
+        _getColumnSpanConfig,
+      }) > 1,
   );
 
   const commonGetPositionArgs = {
@@ -760,10 +861,13 @@ const multiColumnLayout = <T>({
         getPositionsWithMultiColumnItem({
           multiColumnItem: multiColumnItems[i]!,
           itemsToPosition,
+          checkIsFlexibleWidthItem,
+          firstItem,
           heights: acc.heights,
           prevPositions: acc.positions,
           logWhitespace,
           columnCount,
+          responsiveModuleConfigForSecondItem,
           _getColumnSpanConfig,
           _getModulePositioningConfig,
           ...commonGetPositionArgs,